--- a/src/accumulate.jl
+++ b/src/accumulate.jl
@@ -1,13 +1,135 @@
-# Implements the Hillis--Steele algorithm using global memory
-# See algorithm 1 at https://en.wikipedia.org/wiki/Prefix_sum#Parallel_algorithm
-
 # TODO: features
 # - init::Some
 # - CuMatrix
 # - pairwise
 
 # TODO: performance
-# - shared memory / shuffle (see CUDAnative.jl/examples/scan)
+# - shuffle
+# - warp-aggregate atomics
+
+# partial scan of individual thread blocks within a grid
+# work-efficient implementation after Blelloch (1990)
+#
+# number of threads needs to be a power-of-2
+function partial_scan(op::Function, input::CuDeviceVector{T}, output::CuDeviceVector{T},
+                      ::Val{inclusive}=Val(true)) where {T, inclusive}
+    threads = blockDim().x
+    thread = threadIdx().x
+    block = blockIdx().x
+
+    temp = @cuDynamicSharedMem(T, (2*threads,))
+
+    i = (block-1) * threads + thread
+
+    # load input into shared memory
+    @inbounds temp[thread] = if i <= length(input)
+        input[i]
+    else
+        zero(T)
+    end
+
+    # build sum in place up the tree
+    offset = 1
+    d = threads>>1
+    while d > 0
+        sync_threads()
+        @inbounds if thread <= d
+            ai = offset * (2*thread-1)
+            bi = offset * (2*thread)
+            temp[bi] = op(temp[ai], temp[bi])
+        end
+        offset *= 2
+        d >>= 1
+    end
+
+    # clear the last element
+    @inbounds if thread == 1
+        temp[threads] = 0
+    end
+
+    # traverse down tree & build scan
+    d = 1
+    while d < threads
+        offset >>= 1
+        sync_threads()
+        @inbounds if thread <= d
+            ai = offset * (2*thread-1)
+            bi = offset * (2*thread)
+
+            t = temp[ai]
+            temp[ai] = temp[bi]
+            temp[bi] = op(t, temp[bi])
+        end
+        d *= 2
+    end
+
+    sync_threads()
+
+    # write results to device memory
+    @inbounds if i <= length(input)
+        output[i] = if inclusive
+            op(temp[thread], input[i])
+        else
+            temp[thread]
+        end
+    end
+
+    return
+end
+
+# aggregate the result of a partial scan by applying preceding block aggregates
+function aggregate_partial_scan(op::Function, output::CuDeviceVector{T},
+                                aggregates::CuDeviceVector{T}) where {T}
+    threads = blockDim().x
+    thread = threadIdx().x
+    block = blockIdx().x
+
+    i = (block-1) * threads + thread
+
+    @inbounds if block > 1 && i <= length(output)
+        output[i] = op(aggregates[block-1], output[i])
+    end
+
+    return
+end
+
+function Base._accumulate!(f::Function, output::CuVector{T}, input::CuVector{T},
+                           dims::Nothing, init::Nothing) where {T}
+    # determine how many threads we can launch for the scan kernel
+    args = (+, input, output)
+    kernel_args = cudaconvert.(args)
+    kernel_tt = Tuple{Core.Typeof.(kernel_args)...}
+    kernel = cufunction(partial_scan, kernel_tt)
+    kernel_config = launch_configuration(kernel.fun; shmem=(threads)->2*threads*sizeof(T))
+
+    # does that suffice to scan the array in one go?
+    full = nextpow(2, length(input))
+    if full <= kernel_config.threads
+        @cuda name="scan" threads=full shmem=2*full*sizeof(T) partial_scan(+, input, output)
+    else
+        # perform partial scans of smaller thread blocks
+        partial = prevpow(2, kernel_config.threads)
+        blocks = cld(length(input), partial)
+        @cuda threads=partial blocks=blocks shmem=2*partial*sizeof(T) partial_scan(f, input, output)
+
+        # calculate per-block totals
+        aggregates = CuArrays.zeros(T, nextpow(2, blocks))
+        copyto!(aggregates, output[partial:partial:end])
+
+        # scan block totals to get block aggregates
+        accumulate!(f, aggregates, aggregates)
+
+        # apply the block aggregates to the partial scan result
+        # NOTE: we assume that this kernel requires fewer resources than the scan kernel.
+        #       if that does not hold, launch with fewer threads and calculate
+        #       the aggregate block index within the kernel itself.
+        @cuda threads=partial blocks=blocks aggregate_partial_scan(+, output, aggregates)
+
+        CuArrays.unsafe_free!(aggregates)
+    end
+
+    return output
+end
 
 function Base._accumulate!(op::Function, vout::CuVector{T}, v::CuVector, dims::Int,
                            init::Nothing) where {T}
@@ -22,40 +144,7 @@ function Base._accumulate!(op::Function, vout::CuVector{T}, v::CuVector, dims::N
                            init::Nothing) where {T}
     vin = T.(v)  # convert to vector with eltype T
 
-    Δ = 1   # Δ = 2^d
-    n = ceil(Int, log2(length(v)))
-
-    # partial in-place accumulation
-    function kernel(op, vout, vin, Δ)
-        i = threadIdx().x + (blockIdx().x - 1) * blockDim().x
-
-        @inbounds if i <= length(vin)
-            if i > Δ
-                vout[i] = op(vin[i - Δ], vin[i])
-            else
-                vout[i] = vin[i]
-            end
-        end
-
-        return
-    end
-
-    function configurator(kernel)
-        fun = kernel.fun
-        config = launch_configuration(fun)
-        blocks = cld(length(v), config.threads)
-
-        return (threads=config.threads, blocks=blocks)
-    end
-
-    for d in 0:n   # passes through data
-        @cuda config=configurator kernel(op, vout, vin, Δ)
-
-        vin, vout = vout, vin
-        Δ *= 2
-    end
-
-    return vin
+    return Base._accumulate!(op::Function, vout::CuVector{T}, vin::CuVector, dims, init)
 end
 
 Base.accumulate_pairwise!(op, result::CuVector, v::CuVector) = accumulate!(op, result, v)