--- a/src/indexing.jl
+++ b/src/indexing.jl
@@ -26,10 +26,9 @@ function Base.getindex(xs::CuArray{T}, bools::CuArray{Bool}) where {T}
     function kernel(ys::CuDeviceArray{T}, xs::CuDeviceArray{T}, bools, indices)
         i = threadIdx().x + (blockIdx().x - 1) * blockDim().x
 
-        if i <= length(xs) && bools[i]
+        @inbounds if i <= length(xs) && bools[i]
             b = indices[i]   # new position
             ys[b] = xs[i]
-
         end
 
         return
@@ -43,7 +42,7 @@ function Base.getindex(xs::CuArray{T}, bools::CuArray{Bool}) where {T}
         return (threads=config.threads, blocks=blocks)
     end
 
-    @cuda config=configurator kernel(ys, xs, bools, indices)
+    @cuda name="logical_getindex" config=configurator kernel(ys, xs, bools, indices)
   end
 
   unsafe_free!(indices)
@@ -67,10 +66,9 @@ function Base.findall(bools::CuArray{Bool})
         function kernel(ys::CuDeviceArray{Int}, bools, indices)
             i = threadIdx().x + (blockIdx().x - 1) * blockDim().x
 
-            if i <= length(bools) && bools[i]
+            @inbounds if i <= length(bools) && bools[i]
                 b = indices[i]   # new position
                 ys[b] = i
-
             end
 
             return
@@ -84,7 +82,7 @@ function Base.findall(bools::CuArray{Bool})
             return (threads=config.threads, blocks=blocks)
         end
 
-        @cuda config=configurator kernel(ys, bools, indices)
+        @cuda name="findall" config=configurator kernel(ys, bools, indices)
     end
 
     unsafe_free!(indices)